Moved the types and daos out of main into model and daos package.
