+++ main.go
@@ -10,17 +10,17 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
 	"github.com/eddieowens/axon"
 	"github.com/jmoiron/sqlx"
 	_ "github.com/lib/pq"
 	"log"
 	"os"
-	"takeoff.com/matilda/util"
+	"takeoff.com/matilda/daos"
+	"takeoff.com/matilda/model"
 )
 
-func pathsTo(location Location, destination Location) {
+func pathsTo(location model.Location, destination model.Location) {
 	//fmt.Printf("Searching for paths from %v to %v")
 	//seen := map[LocationLabel]bool{}
 	for direction, label := range location.Transitions {
@@ -30,43 +30,6 @@ func pathsTo(location Location, destination Location) {
 	}
 }
 
-type locationArray []Location
-
-type Direction int
-
-const (
-	North Direction = iota
-	South
-	East
-	West
-)
-
-//TODO - get json to serialize using strings below.
-//TODO - look for a way that avoids repeating ourselves.
-func (d Direction) String() string {
-	return [...]string{"North", "South", "East", "West"}[d]
-}
-
-type Plan struct {
-	Locations []Location
-}
-
-type LocationLabel string
-
-type LocationId int
-type Location struct {
-	ID          LocationId
-	Label       LocationLabel
-	Transitions []Transition
-}
-
-type TransitionId int
-type Transition struct {
-	ID 			TransitionId
-	LocationId  LocationId
-	Direction   Direction
-	Destination LocationLabel
-}
 
 var schema = `
 CREATE SEQUENCE location_id_seq;
@@ -83,76 +46,16 @@ CREATE TABLE transition (
     destination text
 );`
 
-func (location Location) String() string {
-	directions := func(location Location) []string {
-		var transitions []string
-		for key, value := range location.Transitions {
-			transitions = append(transitions, fmt.Sprintf("(%v) => %v", key, value))
-		}
-		return transitions
-	}
-	return fmt.Sprintf("%v: %q", location.Label, directions(location))
-}
-
-func (plan *Plan) LoadJson(filename string) error {
-	dat, err := util.Resource(fmt.Sprintf("maps/%s.json", filename))
-	if err == nil {
-		err := json.Unmarshal(dat, &plan)
-		return err
-	} else {
-		return err
-	}
-}
-
-type Crud interface {
-	Insert(sqlx.Tx, Location)
-	Get(sqlx.Tx) []Location
-}
-
-type LocationDao struct {}
-func (LocationDao) Insert(tx sqlx.Tx, location Location) int {
-	stmt, err := tx.PrepareNamed("INSERT INTO location (label) VALUES (:label) RETURNING id")
-	var id int
-	err = stmt.Get(&id, location)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	return id
-}
-func (LocationDao) GetAll(tx sqlx.Tx) []Location {
-	locations := []Location{}
-	tx.Select(&locations, "SELECT * FROM location ORDER BY label ASC")
-	return locations
-}
-func (LocationDao) Map(tx sqlx.Tx) {
-	// Loop through rows using only one struct
-	location := Location{}
-	rows, _ := tx.Queryx("SELECT id, label FROM location")
-	for rows.Next() {
-		err := rows.StructScan(&location)
-		if err != nil {
-			log.Fatalln(err)
-		}
-		fmt.Printf("%#v\n", location)
-	}
-}
-
-type TransitionDao struct {}
-func (TransitionDao) Insert(tx sqlx.Tx, locationId int, transition Transition) {
-	tx.MustExec("INSERT INTO transition (location_id, direction, destination) VALUES ($1, $2, $3)",
-		locationId, transition.Direction, transition.Destination)
-}
-
 type DoSomethingService struct {
 	db sqlx.DB `inject:"db"`
 }
 func (service DoSomethingService) DoSomething() {
-	var locationDao LocationDao
-	var transitionDao TransitionDao
+	var locationDao daos.LocationDao
+	var transitionDao daos.TransitionDao
 
 	// exec the schema or fail; multi-statement Exec behavior varies between
 	// database drivers;  pq will exec them all, sqlite3 won't, ymmv
-	var plan Plan
+	var plan model.Plan
 	err := plan.LoadJson("3x3_floor")
 	if err != nil {
 		fmt.Println(err)
